with test cases for other data types

+++ insertion-sort/insertion-sort.h
@@ -1,5 +1,3 @@
-#include <stdlib.h>
 typedef int (*compare)(void* first, void* second);
-
-void isort(void* base, size_t numberOfElements, size_t elementSize,
+void isort(void* base, int numberOfElements, int elementSize,
 			compare comp);
